apis/transformers: fix tests of Float(32) values
Previously, the expected value for the "Float32" property was assigned
as a constant to a float64 variable, resulting in a stored value of
57.16038. When not assigned as a constant, the stored value became
57.16038131713867 instead of 57.16038.

Both values, when converted to float32, yield 57.16038. However, it is
more accurate to store 57.16038131713867, as that is the got value. The
test did not fail because the comparison occurs after the conversion to
float32.

In the event of a test error, the dumped values would have been
different.

+++ apis/transformers/marshaler_test.go
@@ -172,7 +172,7 @@ var values = []map[string]any{
 		"Uint24":          uint(2880217),
 		"Uint32":          uint(1307298102),
 		"Uint64":          uint(927041163082605),
-		"Float32":         57.16038,
+		"Float32":         float64(float32(57.16038)),
 		"Float64":         18372.36240184391,
 		"Decimal":         decimal.RequireFromString("1752.064"),
 		"DateTime":        time.Date(2023, 10, 17, 9, 34, 25, 836042841, time.UTC),
